DOMA-2247 added getUserDivisionsInfo function
