+++ apps/condo/domains/ticket/access/Ticket.js
@@ -15,7 +15,7 @@ const { getById, find } = require('@core/keystone/schema')
 const { throwAuthenticationError } = require('@condo/domains/common/utils/apolloErrorFormatter')
 const { RESIDENT, STAFF } = require('@condo/domains/user/constants/common')
 const { OrganizationEmployee } = require('@condo/domains/organization/utils/serverSchema')
-const { Division } = require('@condo/domains/division/utils/serverSchema')
+const { Division, getUserDivisionsInfo } = require('@condo/domains/division/utils/serverSchema')
 
 async function canReadTickets ({ authentication: { item: user }, context }) {
     if (!user) return throwAuthenticationError()
@@ -44,24 +44,10 @@ async function canReadTickets ({ authentication: { item: user }, context }) {
         }
     }
 
-    const divisionVisibleLimitedEmployees = await OrganizationEmployee.getAll(context, {
-        user: { id: user.id },
-        role: { isDivisionLimitedVisibility: true },
-        deletedAt: null,
-        isBlocked: false,
-    })
+    const userDivisionsInfo = await getUserDivisionsInfo(context, user.id)
 
-    if (divisionVisibleLimitedEmployees.length > 0) {
-        const organizationsIdsWithEmployeeInDivision = divisionVisibleLimitedEmployees.map(employee => employee.organization.id)
-        const divisionVisibleLimitedEmployeesIds = divisionVisibleLimitedEmployees.map(employee => employee.id)
-        const employeeDivisions = await Division.getAll(context, {
-            OR: [
-                { responsible: { id_in: divisionVisibleLimitedEmployeesIds } },
-                { executors_some: { id_in: divisionVisibleLimitedEmployeesIds } },
-            ],
-        })
-        const divisionsPropertiesIds = uniq(flatten(employeeDivisions.map(division => division.properties))
-            .map(property => property.id))
+    if (userDivisionsInfo) {
+        const { organizationsIdsWithEmployeeInDivision, divisionsPropertiesIds } = userDivisionsInfo
 
         return {
             OR: [
